backend: Move database queries to stored procedures
This commit refactors the existing database queries within the home, report, and search routes to use stored procedures (RPCs) in Supabase. This change improves performance, security, and maintainability by centralizing query logic within the database and reducing the amount of code in the application layer. New functions were written on the supabase side to facilitate this change.